Share registration code between surfaces
